+++ changelog.md
@@ -50,6 +50,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Introduce `_EmpiricalRiskMixin` interface to allow separating deterministic checks and state pre-computation from linear operators
   ([PR](https://github.com/f-dangel/curvlinops/pull/236))
 
+- Generalize sampling output gradients for sequence-valued predictions
+  ([PR](https://github.com/f-dangel/curvlinops/pull/235))
+
 ## [3.0.1] - 2026-01-14
 
 This patch provides major performance improvements for all curvature matrices (by using `torch.func`, and by improving EKFAC's eigenvalue correction) and deprecates Python 3.9 (we now require at least 3.10).